Separating the custom types definition according to https://github.com/rclone/rclone/blob/master/CONTRIBUTING.md#guidelines-for-a-speedy-merge
